main

+++ routes.php
@@ -2,17 +2,17 @@
 
 $router->get('/', 'HomeController@index');
 $router->get('/blog', 'BlogController@index');
-$router->get('/blog/create', 'BlogController@create');
-$router->get('/blog/edit/{id}', 'BlogController@edit');
+$router->get('/blog/create', 'BlogController@create', ['auth']);
+$router->get('/blog/edit/{id}', 'BlogController@edit', ['auth']);
 $router->get('/blog/{id}', 'BlogController@show');
 
-$router->post('/blog', 'BlogController@store');
-$router->delete('/blog/{id}', 'BlogController@destroy');
-$router->put('/blog/{id}', 'BlogController@update');
+$router->post('/blog', 'BlogController@store', ['auth']);
+$router->delete('/blog/{id}', 'BlogController@destroy', ['auth']);
+$router->put('/blog/{id}', 'BlogController@update', ['auth']);
 
-$router->get('/auth/register', 'UserController@create');
-$router->get('/auth/login', 'UserController@login');
+$router->get('/auth/register', 'UserController@create', ['guest']);
+$router->get('/auth/login', 'UserController@login', ['guest']);
 
-$router->post('/auth/register', 'UserController@store');
-$router->post('/auth/logout', 'UserController@logout');
-$router->post('/auth/login', 'UserController@authenticate');
+$router->post('/auth/register', 'UserController@store', ['guest']);
+$router->post('/auth/logout', 'UserController@logout', ['auth']);
+$router->post('/auth/login', 'UserController@authenticate', ['guest']);
